# By elasticthreads
# Via elasticthreads
* 'master' of github.com:ttscoff/nv:
  added numbered list completion
  few more fixes for multi tagging
  improvements to the multiple note tagging
  more on moving cache folder to ~/Library/Caches
  always with the double commit for one fix. more scroller fixes.
  one more fix for the nvalt/10.6 scrollers
  scroller setting in the xib
  little bit more
  truncation
  improved tableview drawing
  fix to a colorscheme bug
  some fixes to how UnifiedCell draws.
  improved grid line drawing in notestableview as well as selection highlighting
  one more slight improvement to the statusview code
  added icon to project file screweup
  deleted old old menu icon. moved the new one
  Improvements to the menubar icon
  Quite a few little fixes
  Improvements to customs scrollers.
  prefs window>> Fonts & Colors >> do a better job of displaying the custom font

Conflicts:
	Notation.xcodeproj/project.pbxproj

+++ BookmarksController.m
@@ -368,7 +368,7 @@ static NSString *BMNoteUUIDStringKey = @"NoteUUIDString";
 }
 
 - (void)doubleClicked:(id)sender {
-	int row = [bookmarksTableView selectedRow];
+	NSInteger row = [bookmarksTableView selectedRow];
 	if (row > -1) [self restoreNoteBookmark:[bookmarks objectAtIndex:row] inBackground:NO];
 }
